add cli variant - converts a document by executing a subprocess of libreoffice using the "--convert-to pdf" argument
